🧹 refactor ms365 provider
