+++ providers/ms365/connection/scopes.go
@@ -2,9 +2,3 @@
 // SPDX-License-Identifier: BUSL-1.1
 
 package connection
-
-const (
-	DefaultMSGraphScope = "https://graph.microsoft.com/.default"
-	TeamsScope          = "48ac35b8-9aa8-4d74-927d-1f4a14a0b239/.default"
-	OutlookScope        = "https://outlook.office.com/.default"
-)
